tests: added integration test for reorg
