+++ packages/daemon/__tests__/integration/scenarios/reorg.balances.ts
@@ -1,0 +1,4 @@
+export default {
+  "HFyF1jYJP9FXfiC3LRqf3q4768TBL1rxbn": 6400,
+  "HMbS5P3NTLQ5oR5TfLNvAkeQ7L8MPn9VM3": 6400
+}
